// frontend/src/client/types.gen.ts
export type ClientPublic = {
    name: string;
    email: string;
    phone_number?: (string | null);
    company_name?: (string | null);
    customer_type: string;
    reg_number?: (string | null);
    billing_address?: (string | null);
    company_id?: (string | null);
    user_id?: (string | null);
    id: string;
    created_at?: (string | null);
};
// ...
export type CompanyPublic = {
    company_name: string;
    currency: string;
    registration_number?: (string | null);
    company_email: string;
    phone_number: string;
    company_url?: (string | null);
    company_address?: (string | null);
    is_active?: boolean;
    document_running_numbers?: {
        [key: string]: (number);
    };
    employee_size?: (string | null);
    business_industry?: (string | null);
    company_type?: (string | null);
    financial_year_end?: (string | null);
    sst_registration_number?: (string | null);
    einvoice_required?: boolean;
    default_document_template_id?: (string | null);
    other_personal_taxable_income?: number;
    id: string;
    user_id?: (string | null);
};
// ...
export type DocumentPublic = {
    company_id?: (string | null);
    user_id?: (string | null);
    docno: string;
    doctype: string;
    client_id?: (string | null);
    date?: string;
    title: string;
    item?: (Array<{
    [key: string]: unknown;
}> | {
    [key: string]: unknown;
});
    price_calculation?: {
        [key: string]: unknown;
    };
    remark?: string;
    status: string;
    validity?: (string | null);
    duedate?: (string | null);
    generated_from_recurring_id?: (string | null);
    stock_deducted?: boolean;
    id: string;
};